api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pillow==11.0.0
numpy==2.2.1
requests==2.32.3
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1
